PrintExtensions.scala moved to scalan.util
