+++ kotlin-backend/src/main/scala/scalan/compilation/kotlin/KotlinFileCodegen.scala
@@ -5,7 +5,7 @@ import java.io.PrintWriter
 import scalan.{Scalan, TypeDesc}
 import scalan.compilation.{IndentLevel, FileCodegen, CodegenConfig}
 import scalan.meta.ScalanAst._
-import scalan.meta.PrintExtensions._
+import scalan.util.PrintExtensions._
 import scalan.meta.{SSymName, ScalanAstTransformers}
 import scalan.primitives.Blocks
